// src/app/models/user/user.model.ts
export class UserModel {
    userName: string = '';
    name: string = '';
    userId: string = '';
    encExpirationDate: string = '';
    currentPassword: string = '';
    password: string = '';
    confirmPassword: string = '';
    firstName: string = '';
    middleName: string = '';
    lastName: string = '';
    salutation: string = '';
    title: string = '';
    phone: string = '';
    phoneExt: string = '';
    afterHoursPhone: string = '';
    afterHoursPhoneExt: string = '';
    cellPhone: string = '';
    fax: string = '';
    faxExt: string = '';
    address1: string = '';
    address2: string = '';
    city: string = '';
    state: string = '';
    zip: string = '';
    country: string = '';
    email: string = '';
    teamId: number = 0;
    landingPage: number = 0;
    officeId: string = '';
    isScheduled: boolean;
    isSurgeon: boolean;
    expiredPasswordAlert: string = '';
    maxLoginAttemptsPassedAlert: string = '';
    activeInd: boolean;
    canAccessPatientFiles: boolean;
    marketingOnly: boolean;
    canPatientMessage: boolean;
    isDietitian: boolean;
    canSaveSettings: boolean;
    securityAccess: number = 0;
}